Re-parse lines containing multiple tweets
Train and test text files have lines that contain multiple tweets and labels. They have been re-parsed so that only a single tweet is on each line. An effect is that the tweets and labels on a multi-tweet line have been appended to the end of the the respective text and label files.